backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0

# ML & Data Processing (optional - uncomment when ready)
# scikit-learn==1.3.2
# pandas==2.1.3
# numpy==1.26.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0



